nshipping/requirements.txt
Django==1.8.4
Pillow==2.5.3
Unipath==1.1
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-braces==1.8.1
django-model-utils==2.3.1
django-toolbelt==0.0.1
djangorestframework==3.3.2
gunicorn==19.3.0
html5lib==0.9999999
git+https://github.com/bluehawkarthur/pisa.git
psycopg2==2.6.1
reportlab==3.2.0
six==1.10.0
static3==0.6.1
wheel==0.26.0
wsgiref==0.1.2
